requirements.txt
pandas==3.0.1
openpyxl==3.1.5
streamlit==1.55.0
groq==1.0.0
plotly==6.6.0
python-dotenv==1.2.2
